requirements.txt
pyqt5
pillow
configparser
pywin32
pyinstaller
pysmb